requirements.txt
absl-py==2.3.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
flatbuffers==25.2.10
gast==0.6.0
git-filter-repo==2.47.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.73.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.1.0
tf_keras==2.18.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.14.1
